Added @, if, and comma to help parsing
Without the @, the {username} looks odd and doesn't automatically link using GitHub. Also, given that the `or` used here separates two conditional clauses and not noun clauses, it is better to have an `if` next to the second clause to help humans with parsing the sentence. The oxford comma makes the `please` stand out as the head of its own clause.